requirements.txt
# Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tenacity==8.2.3
cachetools==5.3.2

# Data & API
pandas==2.1.0
requests==2.31.0
google-api-python-client==2.100.0

# Other
beautifulsoup4==4.12.2
instaloader==4.10.3
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.13.0
